refactor: adopt clean architecture structure
