+++ SunatScraper.Api/SunatScraper.Api.csproj
@@ -6,7 +6,8 @@
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
   <ItemGroup>
-    <ProjectReference Include="..\SunatScraper.Core\SunatScraper.Core.csproj" />
+    <ProjectReference Include="..\SunatScraper.Domain\SunatScraper.Domain.csproj" />
+    <ProjectReference Include="..\SunatScraper.Infrastructure\SunatScraper.Infrastructure.csproj" />
     <PackageReference Include="Serilog.AspNetCore" Version="7.*" />
   </ItemGroup>
 </Project>
