+++ SunatScraper.Domain/ISunatClient.cs
@@ -1,0 +1,11 @@
+namespace SunatScraper.Domain;
+using SunatScraper.Domain.Models;
+
+public interface ISunatClient
+{
+    Task<RucInfo> ByRucAsync(string ruc);
+    Task<RucInfo> ByDocumentoAsync(string tipo, string numero);
+    Task<IReadOnlyList<SearchResultItem>> SearchDocumentoAsync(string tipo, string numero);
+    Task<IReadOnlyList<SearchResultItem>> SearchRazonAsync(string query);
+    Task<RucInfo> ByRazonAsync(string query);
+}
