+++ SunatScraper.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SunatScraper.Api", "SunatScraper.Api\SunatScraper.Api.csproj", "{5A6F4E3F-C422-4A3D-A618-15688EE837DB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SunatScraper.Core", "SunatScraper.Core\SunatScraper.Core.csproj", "{F05A6242-D14B-4423-8218-B6375D3E888B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SunatScraper.Infrastructure", "SunatScraper.Infrastructure\SunatScraper.Infrastructure.csproj", "{F05A6242-D14B-4423-8218-B6375D3E888B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SunatScraper.Domain", "SunatScraper.Domain\SunatScraper.Domain.csproj", "{D54A8E3C-1234-47E0-BAD1-111111111111}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SunatScraper.Grpc", "SunatScraper.Grpc\SunatScraper.Grpc.csproj", "{3210FDCF-8A00-4279-AA6F-0806A13AAFBC}"
 EndProject
@@ -44,6 +46,18 @@ Global
 		{F05A6242-D14B-4423-8218-B6375D3E888B}.Release|x86.ActiveCfg = Release|Any CPU
 		{F05A6242-D14B-4423-8218-B6375D3E888B}.Release|x86.Build.0 = Release|Any CPU
 		{3210FDCF-8A00-4279-AA6F-0806A13AAFBC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {D54A8E3C-1234-47E0-BAD1-111111111111}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {D54A8E3C-1234-47E0-BAD1-111111111111}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {D54A8E3C-1234-47E0-BAD1-111111111111}.Debug|x64.ActiveCfg = Debug|Any CPU
+                {D54A8E3C-1234-47E0-BAD1-111111111111}.Debug|x64.Build.0 = Debug|Any CPU
+                {D54A8E3C-1234-47E0-BAD1-111111111111}.Debug|x86.ActiveCfg = Debug|Any CPU
+                {D54A8E3C-1234-47E0-BAD1-111111111111}.Debug|x86.Build.0 = Debug|Any CPU
+                {D54A8E3C-1234-47E0-BAD1-111111111111}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {D54A8E3C-1234-47E0-BAD1-111111111111}.Release|Any CPU.Build.0 = Release|Any CPU
+                {D54A8E3C-1234-47E0-BAD1-111111111111}.Release|x64.ActiveCfg = Release|Any CPU
+                {D54A8E3C-1234-47E0-BAD1-111111111111}.Release|x64.Build.0 = Release|Any CPU
+                {D54A8E3C-1234-47E0-BAD1-111111111111}.Release|x86.ActiveCfg = Release|Any CPU
+                {D54A8E3C-1234-47E0-BAD1-111111111111}.Release|x86.Build.0 = Release|Any CPU
 		{3210FDCF-8A00-4279-AA6F-0806A13AAFBC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3210FDCF-8A00-4279-AA6F-0806A13AAFBC}.Debug|x64.ActiveCfg = Debug|Any CPU
 		{3210FDCF-8A00-4279-AA6F-0806A13AAFBC}.Debug|x64.Build.0 = Debug|Any CPU
